server/requirements.txt
uvicorn
litestar
litestar[standard]
pydantic
pydantic[email]
passlib
redis
pydantic-settings
python-dotenv
alembic
